refactor: Extract sync logic to own mod
