+++ coordinator/src/synchronise/mod.rs
@@ -1,0 +1,5 @@
+mod block_streams;
+mod executors;
+
+pub use block_streams::synchronise_block_streams;
+pub use executors::synchronise_executors;
